Re-write background and content scripts
Also don't remove injection scripts for any build
