+++ package.json
@@ -3,5 +3,13 @@
     "adm-zip": "^0.5.9",
     "es-main": "^1.2.0"
   },
-  "type": "module"
+  "type": "module",
+  "devDependencies": {
+    "eslint": "^8.36.0",
+    "eslint-config-prettier": "^8.8.0",
+    "prettier": "^2.8.7"
+  },
+  "scripts": {
+    "lint": "eslint"
+  }
 }
